version 1.0 GPS Tracking About Page Tarkie Flavor Overtime Cancel Process Send Backup SSZipArchive pod HTTP POST FILE Layout margins to constraint transition Scrollview library UIPageViewController swipe Keyboard swipe View adjust to keyboard IOS 10 compatibility Camera flash off by default CircularProgressBar library ContainerView

+++ tarkie/widgets/collectionView/PageBarCollectionView.h
@@ -3,7 +3,7 @@
 @protocol PageBarDelegate
 @optional
 
-- (void)onPageBarSelect:(int)page;
+- (void)onPageBarSelect:(long)page;
 
 @end
